ATM.cpp
"miniATM" is a project designed to simulate the functionality of an ATM (Automated Teller Machine). Its features include pin generation, pin change, withdrawal of money, and deposit of money. Users can generate a PIN for their account, change it as needed, withdraw funds, and deposit money into their account. This project aims to provide a convenient and secure way for users to manage their finances in a simulated ATM environment. The GitHub repository contains the source code and documentation for the project, enabling developers to explore, contribute, and utilize the miniATM system.


------------------------------------------------------------
How to use this project
1. Pin generate
2. Change pin
After that you can do anything

Thank you for seeing my project